+++ sdk/resourcemanager/maps/armmaps/CHANGELOG.md
@@ -1,5 +1,22 @@
 # Release History
 
+## 2.0.0-beta.1 (2023-11-14)
+### Breaking Changes
+
+- Function `timeRFC3339.MarshalText` has been removed
+- Function `*timeRFC3339.Parse` has been removed
+- Function `*timeRFC3339.UnmarshalText` has been removed
+
+### Features Added
+
+- New function `dateTimeRFC3339.MarshalText() ([]byte, error)`
+- New function `*dateTimeRFC3339.Parse(string) error`
+- New function `*dateTimeRFC3339.UnmarshalText([]byte) error`
+- New struct `LocationsItem`
+- New field `Locations` in struct `AccountProperties`
+- New field `ConsumedStorageUnitSizeInBytes`, `TotalStorageUnitSizeInBytes` in struct `CreatorProperties`
+
+
 ## 1.0.0 (2023-08-25)
 ### Breaking Changes